Add global error handling for tan-query
